hw10 style fix

+++ src/main/java/edu/hw10/task2/entity/FibCalculator.java
@@ -5,5 +5,5 @@ import edu.hw10.task2.annotation.Cache;
 public interface FibCalculator {
 
     @Cache(persist = true)
-    long fib(int number);
+    int fib(int number);
 }
